Added two migration scripts.
Created two new migration scripts in order for alembic to detect the change of names for the genres attrubute. This was done in order to add the property db.ARRAY() to the Artist Model attribute. Alembic could not auto detect the changes to the added property arguement for the attribute. Hence, the justification for this process.